+++ CHANGELOG.md
@@ -4,7 +4,11 @@
 
 - Compatibility with Bitcoin Core v0.21
 
+- Support for Signet
+
 - New `create_wallet_if_missing` option to ease the creation of a designated bitcoind wallet (#76)
+
+- Upgrade to rust-bitcoin v0.26 & rust-miniscript v5.0
 
 ## 0.2.1 - 2021-01-14